完成计算机的缩进，进行删除账单

+++ app/src/main/java/com/phone/konka/accountingbook/Bean/MonthDetailBean.java
@@ -40,7 +40,7 @@ public class MonthDetailBean {
 
     private List<DayDetailBean> dayList;
 
-    public double getYear() {
+    public int getYear() {
         return year;
     }
